Move the codemirror editors to their own Component folder
