+++ studio/src/components/Timeline/DetailsList/index.ts
@@ -1,9 +1,3 @@
 export { CollapsibleKeyValueTableV2 } from "./KeyValueTableV2";
-export {
-  CodeMirrorSqlEditor,
-  CodeMirrorJsonEditor,
-  CodeMirrorTypescriptEditor,
-  CodeMirrorInput,
-} from "./CodeMirrorEditor";
 export { TimelineListDetails } from "./TimelineDetailsList";
 export { BodyViewerV2 } from "./BodyViewerV2";
