+++ studio/src/components/Timeline/index.ts
@@ -4,9 +4,6 @@ export { RequestMethod, SubSectionHeading } from "./shared";
 export { useAsWaterfall } from "./hooks";
 export {
   BodyViewerV2,
-  CodeMirrorJsonEditor,
-  CodeMirrorSqlEditor,
-  CodeMirrorTypescriptEditor,
   CollapsibleKeyValueTableV2,
   TimelineListDetails,
 } from "./DetailsList";
